Just classes layout in packages refactored and some javadoc warnings fixed.
