+++ java/org/moskito/control/core/Application.java
@@ -1,5 +1,7 @@
 package org.moskito.control.core;
 
+import org.moskito.control.core.chart.Chart;
+
 import java.util.LinkedList;
 import java.util.List;
