+++ java/org/moskito/control/core/HealthColor.java
@@ -47,7 +47,7 @@ public enum HealthColor {
      * Returns appropriate threshold color by {@link ThresholdStatus} value.
      * Remains for having single color set through MoSKito Control.
      *
-     * @see org.moskito.control.connectors.V1Parser;
+     * @see org.moskito.control.connectors.parsers.V1Parser ;
      * @param status {@link ThresholdStatus}.
      * @return {@link ThresholdStatus}.
      */
