+++ java/org/moskito/control/core/chart/Chart.java
@@ -1,4 +1,7 @@
-package org.moskito.control.core;
+package org.moskito.control.core.chart;
+
+import org.moskito.control.core.Application;
+import org.moskito.control.core.accumulator.AccumulatorDataItem;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
