+++ java/org/moskito/control/core/history/StatusUpdateHistoryRepository.java
@@ -2,8 +2,8 @@ package org.moskito.control.core.history;
 
 import org.moskito.control.core.Application;
 import org.moskito.control.core.ApplicationRepository;
-import org.moskito.control.core.StatusChangeEvent;
-import org.moskito.control.core.StatusChangeListener;
+import org.moskito.control.core.status.StatusChangeEvent;
+import org.moskito.control.core.status.StatusChangeListener;
 
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
