+++ java/org/moskito/control/core/updater/ChartDataUpdater.java
@@ -4,12 +4,12 @@ import org.moskito.control.config.ComponentConfig;
 import org.moskito.control.config.MoskitoControlConfiguration;
 import org.moskito.control.config.UpdaterConfig;
 import org.moskito.control.connectors.Connector;
-import org.moskito.control.connectors.ConnectorAccumulatorResponse;
+import org.moskito.control.connectors.response.ConnectorAccumulatorResponse;
 import org.moskito.control.connectors.ConnectorFactory;
-import org.moskito.control.core.AccumulatorDataItem;
+import org.moskito.control.core.accumulator.AccumulatorDataItem;
 import org.moskito.control.core.Application;
 import org.moskito.control.core.ApplicationRepository;
-import org.moskito.control.core.Chart;
+import org.moskito.control.core.chart.Chart;
 import org.moskito.control.core.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
