+++ java/org/moskito/control/ui/action/MainViewAction.java
@@ -10,13 +10,13 @@ import net.anotheria.util.TimeUnit;
 import net.anotheria.util.sorter.DummySortType;
 import net.anotheria.util.sorter.StaticQuickSorter;
 import org.moskito.control.config.MoskitoControlConfiguration;
-import org.moskito.control.core.AccumulatorDataItem;
+import org.moskito.control.core.accumulator.AccumulatorDataItem;
 import org.moskito.control.core.Application;
 import org.moskito.control.core.ApplicationRepository;
-import org.moskito.control.core.Chart;
-import org.moskito.control.core.ChartLine;
+import org.moskito.control.core.chart.Chart;
+import org.moskito.control.core.chart.ChartLine;
 import org.moskito.control.core.Component;
-import org.moskito.control.core.ThresholdDataItem;
+import org.moskito.control.core.threshold.ThresholdDataItem;
 import org.moskito.control.core.history.StatusUpdateHistoryItem;
 import org.moskito.control.core.history.StatusUpdateHistoryRepository;
 import org.moskito.control.core.notification.StatusChangeMailNotifier;
